Fixed typos and bugs with the organic and aqueous phases
Added the height of the mixer-settler units to be 8 ft on all units. Fixed some typos and removed test print statements.

The bug for the aqueous phase was caused by the density of the feed being added with the density of the raffinate (effectively doubling the density), this needs to be averaged instead.

The bug for the organic phase was caused by adding the raffinate mass flow rate to the organic phase (raffinate is the aqueous phase).